assistant/requirements.txt
streamlit
opencv-python
ultralytics
sounddevice
scipy
python-dotenv
openai
pyttsx3
